backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
langchain==0.3.1
langgraph==0.2.4
langchain-openai==0.1.17
langchain-community==0.3.1
numpy==1.26.4
pydantic==2.9.2
python-multipart==0.0.7
python-jose[cryptography]==3.3.0
redis==5.1.0
openpyxl==3.1.5
pypdf==4.3.1
docx2txt==0.8
unstructured[all]==0.14.0